Allow OGUSA submission for all years
- add OGUSA tests with mock assets
- extract ogusa submit functions into Compute class
