+++ templates/dynamic/dynamic_input_form.html
@@ -52,7 +52,7 @@
               </ul>
               <div class="sidebar-button">
                 <a href="#" ></a>
-                <input id="tax-submit" class="btn btn-secondary btn-block btn-animate disabled" disabled type="submit" value="Start Dynamic Simulation!">
+                <input id="tax-submit" class="btn btn-secondary btn-block btn-animate {{ is_disabled }}" {{ is_disabled }} type="submit" value="Start Dynamic Simulation!">
                 <div class="construction text-center">
                   <div class="progress progress-striped" style="width: 100%">
                     <div class="progress-bar active progress-bar-custom" role="progressbar" aria-valuenow="45" aria-valuemin="0" aria-valuemax="100" style="width: 100%"></div>
