+++ templates/dynamic/landing.html
@@ -67,7 +67,6 @@
                       {% flatblock "dynamic_behavioral_blurb" %}
                     </td>
                 </tr>
-                {% if is_authenticated %}
                 <tr padding="15px">
                     <td>
                     <a href="/dynamic/ogusa/{{pk}}/?start_year={{start_year}}" class="text-white btn btn-secondary">Overlapping Generations Simulation</a>
@@ -81,7 +80,6 @@
                       {% flatblock "dynamic_olg_blurb" %}
                     </td>
                 </tr>
-                {% endif %}
                 <tr padding="15px">
                     <td padding-bottom="1em">
                      <a href="/dynamic/macro/{{pk}}/?start_year={{start_year}}" class="text-white btn btn-secondary">Macro Elasticities Simulation</a>
