+++ webapp/apps/dynamic/helpers.py
@@ -204,145 +204,6 @@ def filter_ogusa_only(user_values):
             user_values[k] = float(v)
 
     return user_values
- 
-def submit_ogusa_calculation(mods, first_budget_year, microsim_data):
-    print "mods is ", mods
-    ogusa_mods = filter_ogusa_only(mods)
-    microsim_params = package_up_vars(microsim_data, first_budget_year)
-    print "submit dynamic work"
-    print "ogusa_mods is ", ogusa_mods
-
-    hostnames = OGUSA_WORKERS
-
-    DEFAULT_PARAMS = {
-        'callback': "http://{}/dynamic/dynamic_finished".format(CALLBACK_HOSTNAME),
-    }
-
-    data = {}
-    data['ogusa_params'] = json.dumps(ogusa_mods)
-    microsim_mods = {first_budget_year:microsim_params}
-    data['user_mods'] = json.dumps(microsim_mods)
-    job_ids = []
-    guids = []
-    hostname_idx = OGUSA_WORKER_IDX
-    submitted = False
-    registered = False
-    attempts = 0
-    while not submitted:
-        theurl = "http://{hn}/ogusa_start_job".format(hn=hostnames[hostname_idx])
-        try:
-            response = requests.post(theurl, data=data, timeout=TIMEOUT_IN_SECONDS)
-            if response.status_code == 200:
-                print "submitted: ", hostnames[hostname_idx]
-                submitted = True
-                resp_data = json.loads(response.text)
-                job_ids.append((resp_data['job_id'], hostnames[hostname_idx]))
-                guids.append((resp_data['job_id'], resp_data.get('guid', 'None')))
-            else:
-                print "FAILED: ", hostnames[hostname_idx]
-                attempts += 1
-        except Timeout:
-            print "Couldn't submit to: ", hostnames[hostname_idx]
-            increment_ogusa_worker()
-            attempts += 1
-        except RequestException as re:
-            print "Something unexpected happened: ", re
-            increment_ogusa_worker()
-            attempts += 1
-        if attempts > MAX_ATTEMPTS_SUBMIT_JOB:
-            print "Exceeded max attempts. Bailing out."
-            increment_ogusa_worker()
-            raise IOError()
-
-    params = DEFAULT_PARAMS.copy()
-    params['job_id'] = job_ids[0]
-    reg_url = "http://" + hostnames[hostname_idx] + "/register_job"
-    register = requests.post(reg_url, data=params)
-
-    while not registered:
-        reg_url = "http://{hn}/register_job".format(hn=hostnames[hostname_idx])
-        try:
-            params = DEFAULT_PARAMS.copy()
-            params['job_id'] = job_ids[0][0]
-            reg_url = "http://" + hostnames[hostname_idx] + "/register_job"
-
-            register = requests.post(reg_url, data=params, timeout=TIMEOUT_IN_SECONDS)
-            if response.status_code == 200:
-                print "registered: ", hostnames[hostname_idx]
-                registered = True
-            else:
-                print "FAILED: ", hostnames[hostname_idx]
-                attempts += 1
-        except Timeout:
-            print "Couldn't submit to: ", hostnames[hostname_idx]
-            attempts += 1
-        except RequestException as re:
-            print "Something unexpected happened: ", re
-            attempts += 1
-        if attempts > MAX_ATTEMPTS_SUBMIT_JOB:
-            print "Exceeded max attempts. Bailing out."
-            raise IOError()
-
-    # We increment upon exceptions to submit, but once we have submitted and
-    # registered, increment again to move to the next OGUSA worker node
-    increment_ogusa_worker()
-    return job_ids, guids
-
-
-# Might not be needed because this would be handled on the worker node side
-def ogusa_results_ready(job_ids):
-    jobs_done = [False] * len(job_ids)
-    for idx, id_hostname in enumerate(job_ids):
-        id_, hostname = id_hostname
-        result_url = "http://{hn}/dropq_query_result".format(hn=hostname)
-        job_response = requests.get(result_url, params={'job_id': id_})
-        if job_response.status_code == 200: # Valid response
-            rep = job_response.text
-            if rep == 'YES':
-                jobs_done[idx] = True
-                print "got one!: ", id_
-
-    return all(jobs_done)
-
-
-def ogusa_get_results(job_ids, status):
-    '''
-    job_ids = celery ID and hostname of job
-    status = either "SUCCESS" or "FAILURE"
-    '''
-    id_hostname  = job_ids[0]
-    id_, hostname = id_hostname
-    result_url = "http://{hn}/dropq_get_result".format(hn=hostname)
-    job_response = requests.get(result_url, params={'job_id':id_})
-    if job_response.status_code == 200: # Valid response
-        if status == "SUCCESS":
-            response = job_response.json()
-            df_ogusa = {}
-            df_ogusa.update(response['df_ogusa'])
-            results = {'df_ogusa': df_ogusa}
-        elif status == "FAILURE":
-            results = {'job_fail': job_response.text}
-        else:
-            raise ValueError("only know 'SUCCESS' or 'FAILURE' status")
-    else:
-        msg = "Don't know how to handle response: {0}"
-        msg = msg.format(job_response.status_code)
-        raise IOError(msg)
-
-    if ENFORCE_REMOTE_VERSION_CHECK:
-        versions = [r.get('ogusa_version', None) for r in ans]
-        if not all([ver==ogusa_version for ver in versions]):
-            msg ="Got different taxcalc versions from workers. Bailing out"
-            print msg
-            raise IOError(msg)
-        versions = [r.get('dropq_version', None) for r in ans]
-        if not all([same_version(ver, dropq_version) for ver in versions]):
-            msg ="Got different dropq versions from workers. Bailing out"
-            print msg
-            raise IOError(msg)
-
-
-    return results
 
 
 def job_submitted(email_addr, model):
