+++ webapp/apps/dynamic/tests/ogusa_results_0.json
@@ -1,0 +1,1 @@
+{"dropq_version": "0.6.d73f59", "ogusa_version": "0.5.3.88ff22", "df_ogusa": {"GDP": ["0.002", "0.001", "0.000", "0.000", "-0.000", "-0.000", "0.000", "0.000", "0.000", "0.000", "0.000", "0.001"], "Consumption": ["0.010", "0.005", "0.004", "0.003", "0.003", "0.003", "0.003", "0.003", "0.003", "0.003", "0.004", "0.001"], "Interest Rates": ["0.003", "0.005", "0.007", "0.007", "0.009", "0.009", "0.010", "0.010", "0.010", "0.010", "0.008", "0.010"], "Total Taxes": ["-0.024", "-0.026", "-0.026", "-0.026", "-0.025", "-0.026", "-0.027", "-0.027", "-0.027", "-0.027", "-0.026", "-0.025"], "Wages": ["-0.001", "-0.002", "-0.003", "-0.003", "-0.003", "-0.003", "-0.003", "-0.003", "-0.003", "-0.003", "-0.003", "-0.004"], "Investment": ["-0.030", "-0.014", "-0.008", "-0.009", "-0.007", "-0.007", "-0.007", "-0.007", "-0.007", "-0.007", "-0.010", "-0.007"], "Hours Worked": ["0.002", "0.003", "0.003", "0.003", "0.003", "0.003", "0.003", "0.004", "0.004", "0.004", "0.003", "0.005"]}, "taxcalc_version": "0.6.10d462"}
