+++ webapp/apps/taxbrain/tests/test_views.py
@@ -10,6 +10,8 @@ from ..compute import DropqCompute, MockCompute, MockFailedCompute
 import taxcalc
 from taxcalc import Policy
 
+START_YEAR = 2016
+
 class TaxBrainViewsTests(TestCase):
     ''' Test the views of this app. '''
 
@@ -38,7 +40,7 @@ class TaxBrainViewsTests(TestCase):
                 u'ID_BenefitSurtax_Switch_4': [u'True'],
                 u'ID_BenefitSurtax_Switch_6': [u'True'],
                 u'has_errors': [u'False'], u'II_em': [u'4333'],
-                u'start_year': u'2016', 'csrfmiddlewaretoken':'abc123'}
+                u'start_year': unicode(START_YEAR), 'csrfmiddlewaretoken':'abc123'}
 
         response = self.client.post('/taxbrain/', data)
         # Check that redirect happens
@@ -62,7 +64,7 @@ class TaxBrainViewsTests(TestCase):
                 u'ID_BenefitSurtax_Switch_4': [u'True'],
                 u'ID_BenefitSurtax_Switch_6': [u'True'],
                 u'has_errors': [u'False'], u'BE_inc': [u'0.1'],
-                u'start_year': u'2016'}
+                u'start_year': unicode(START_YEAR)}
 
         response = self.client.post('/taxbrain/', data)
         # Check that we get a 400
@@ -83,7 +85,7 @@ class TaxBrainViewsTests(TestCase):
                 u'ID_BenefitSurtax_Switch_4': [u'True'],
                 u'ID_BenefitSurtax_Switch_6': [u'True'],
                 u'has_errors': [u'False'], u'II_em': [u'4333'],
-                u'start_year': u'2016', 'csrfmiddlewaretoken':'abc123'}
+                u'start_year': unicode(START_YEAR), 'csrfmiddlewaretoken':'abc123'}
 
         response = self.client.post('/taxbrain/', data)
         # Check that redirect happens
